fix:use lowercase to rename all file in views/
